refactoring

+++ src/business_logic/business_logic.cpp
@@ -7,28 +7,37 @@
  */
 
 #include <algorithm>
+#include <iostream>
 #include <set>
 
+#include "../communicate/buffer_tools.h"
 #include "business_logic.h"
 #include "string_extensions.h"
 
 namespace business_logic
 {
 
-std::string business_logic::calculate(const std::string &str)
+bool business_logic::calculate(communicate::buffer& b, int type)
 {
+    std::string str = communicate::buffer_get(b);
+
+    if(str.size() == 0)
+        return false;
+
+    std::cout << (type == 1 ? "udp <<< " : "tcp <<< ") << str << std::endl;
+
     auto strings = string_extensions::split(
                 string_extensions::remove_all(
                     string_extensions::remove_all(str, ','), '.'), ' ');
 
     std::multiset<int> nums;
 
-    for_each (strings.begin(), strings.end(), [&](const std::string &s)
+    for_each(strings.begin(), strings.end(), [&](const std::string& s)
     {
         if(s.size() == 0)
             return ;
 
-        if(std::any_of(s.begin(), s.end(), [](const char &c)
+        if(std::any_of(s.begin(), s.end(), [](const char& c)
         {
             if(c < 0x30 || c > 0x39)
                 return true;
@@ -41,17 +50,16 @@ std::string business_logic::calculate(const std::string &str)
         {
             nums.insert(std::stoi(s));
         }
-        catch (...) { }
+        catch(...) { }
     });
 
     int sum = 0;
 
     std::string result;
 
-    for_each (nums.begin(), nums.end(), [&](const int &i)
+    for_each(nums.begin(), nums.end(), [&](const int& i)
     {
         result += std::to_string(i) + " ";
-
         sum += i;
     });
 
@@ -59,8 +67,22 @@ std::string business_logic::calculate(const std::string &str)
         result.resize(result.size() - 1);
 
     result += '\n' + std::to_string(sum) + '\0';
+    communicate::buffer_set(b, result);
+    std::cout << (type == 1 ? "udp >>> " : "tcp >>> ") << str << std::endl;
+    return true;
+}
 
-    return result;
+void business_logic::input(communicate::buffer& b)
+{
+    std::string s;
+    std::getline(std::cin, s);
+    buffer_set(b, s);
+}
+
+void business_logic::output(const communicate::buffer& b, ssize_t number_of_bytes_resived)
+{
+    std::cout << "bytes recieved: " << number_of_bytes_resived << std::endl;
+    std::cout << "bytes: " << std::string((char*)b.data, number_of_bytes_resived) << std::endl;
 }
 
 }
